+++ examples/data-sources/iosxr_l2vpn_pw_class/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_l2vpn_pw_class" "example" {
+  name = "PWC1"
+}