+++ examples/resources/iosxr_l2vpn_pw_class/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxr_l2vpn_pw_class.example "Cisco-IOS-XR-um-l2vpn-cfg:/l2vpn/pw-classes/pw-class[pw-class-name=PWC1]"